test: reorganize App.test.ts into focused modules
- Broke down 3,151 line App.test.ts into 11 organized integration test modules
- Created focused test files: file-selection, file-comparison, keyboard-navigation,
  copy-operations, menu-settings, minimap-interaction, scroll-sync, save-operations,
  quit-dialog, edge-cases, and startup tests
- Added 4 new edge case tests: unusual file extensions, long file names/paths,
  special characters in names, and different line endings (CRLF vs LF)
- Removed redundant skipped test from startup module
- All 47 integration tests now pass with improved organization and maintainability
