+++ frontend/src/App.test.ts
@@ -1,3151 +1,19 @@
-import { fireEvent, render, screen, waitFor } from "@testing-library/svelte";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import App from "./App.svelte";
-
-// Mock the Wails runtime
-vi.mock("../wailsjs/runtime/runtime.js", () => ({
-	EventsOn: vi.fn(),
-	EventsOff: vi.fn(),
-}));
-
-// Mock the Wails App functions
-vi.mock("../wailsjs/go/main/App.js", () => ({
-	CompareFiles: vi.fn(),
-	GetInitialFiles: vi.fn(),
-	GetMinimapVisible: vi.fn(),
-	SetMinimapVisible: vi.fn(),
-	HasUnsavedChanges: vi.fn(),
-	SaveChanges: vi.fn(),
-	CopyToFile: vi.fn(),
-	DiscardAllChanges: vi.fn(),
-	QuitWithoutSaving: vi.fn(),
-	SaveSelectedFilesAndQuit: vi.fn(),
-	SelectFile: vi.fn(),
-}));
-
-import {
-	CompareFiles,
-	CopyToFile,
-	GetInitialFiles,
-	GetMinimapVisible,
-	HasUnsavedChanges,
-	SetMinimapVisible,
-	SaveChanges,
-	SelectFile,
-} from "../wailsjs/go/main/App.js";
-
-import { EventsOn } from "../wailsjs/runtime/runtime.js";
-
-describe("App Component - File Selection", () => {
-	beforeEach(() => {
-		vi.clearAllMocks();
-
-		// Default mocks
-		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
-		vi.mocked(GetMinimapVisible).mockResolvedValue(true);
-	});
-
-	it("should handle file selection and update UI accordingly", async () => {
-		// Mock different file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/script.js")
-			.mockResolvedValueOnce("/path/to/styles.css")
-			.mockResolvedValueOnce("") // Cancel scenario
-			.mockResolvedValueOnce("/path/to/index.html");
-
-		const { container } = render(App);
-
-		// Get all file buttons and compare button
-		const fileButtons = container.querySelectorAll(".file-btn");
-		const leftFileButton = fileButtons[0] as HTMLButtonElement;
-		const rightFileButton = fileButtons[1] as HTMLButtonElement;
-		const compareButton = container.querySelector(".compare-btn") as HTMLButtonElement;
-
-		expect(leftFileButton).toBeTruthy();
-		expect(rightFileButton).toBeTruthy();
-		expect(compareButton).toBeTruthy();
-
-		// Initially, compare button should be disabled
-		expect(compareButton.disabled).toBe(true);
-		expect(leftFileButton.textContent).toContain("Select left file...");
-		expect(rightFileButton.textContent).toContain("Select right file...");
-
-		// Select left file
-		await fireEvent.click(leftFileButton);
-
-		// Verify left file selection
-		await waitFor(() => {
-			expect(SelectFile).toHaveBeenCalledTimes(1);
-			expect(leftFileButton.textContent).toContain("script.js");
-			
-			// Check icon update
-			const leftIcon = leftFileButton.querySelector(".file-icon");
-			expect(leftIcon?.getAttribute("title")).toBe("JavaScript");
-			
-			// Compare button should still be disabled (only one file selected)
-			expect(compareButton.disabled).toBe(true);
-		});
-
-		// Select right file
-		await fireEvent.click(rightFileButton);
-
-		// Verify right file selection
-		await waitFor(() => {
-			expect(SelectFile).toHaveBeenCalledTimes(2);
-			expect(rightFileButton.textContent).toContain("styles.css");
-			
-			// Check icon update
-			const rightIcon = rightFileButton.querySelector(".file-icon");
-			expect(rightIcon?.getAttribute("title")).toBe("CSS");
-			
-			// Compare button should now be enabled
-			expect(compareButton.disabled).toBe(false);
-		});
-
-		// Test cancel scenario - click left file again but cancel
-		await fireEvent.click(leftFileButton);
-
-		// Verify file remains unchanged when cancelled
-		await waitFor(() => {
-			expect(SelectFile).toHaveBeenCalledTimes(3);
-			// File should remain the same
-			expect(leftFileButton.textContent).toContain("script.js");
-			expect(compareButton.disabled).toBe(false);
-		});
-
-		// Test file replacement - select different right file
-		await fireEvent.click(rightFileButton);
-
-		// Verify file replacement
-		await waitFor(() => {
-			expect(SelectFile).toHaveBeenCalledTimes(4);
-			expect(rightFileButton.textContent).toContain("index.html");
-			
-			// Check icon update for HTML file
-			const rightIcon = rightFileButton.querySelector(".file-icon");
-			expect(rightIcon?.getAttribute("title")).toBe("HTML");
-			
-			// Compare button should remain enabled
-			expect(compareButton.disabled).toBe(false);
-		});
-
-		// Verify both files are properly set
-		const leftFileName = leftFileButton.querySelector(".file-name");
-		const rightFileName = rightFileButton.querySelector(".file-name");
-		expect(leftFileName?.textContent).toBe("script.js");
-		expect(rightFileName?.textContent).toBe("index.html");
-	});
-
-});
-
-describe("App Component - File Comparison", () => {
-	beforeEach(() => {
-		vi.clearAllMocks();
-
-		// Default mocks
-		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
-		vi.mocked(GetMinimapVisible).mockResolvedValue(true);
-	});
-
-	it("should show activity indicator during comparison", async () => {
-		// Mock file selection
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles to have a delay to see the loading state
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockImplementation(
-			() =>
-				new Promise((resolve) =>
-					setTimeout(() => resolve(mockDiffResult), 200),
-				),
-		);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		// Click compare
-		await fireEvent.click(compareButton!);
-
-		// Should immediately show "Comparing..."
-		expect(compareButton?.textContent).toBe("Comparing...");
-		expect(compareButton).toHaveProperty("disabled", true);
-
-		// Wait for comparison to complete
-		await waitFor(
-			() => {
-				expect(compareButton?.textContent).toBe("Compare");
-				// Button should remain disabled after comparison completes
-				expect(compareButton).toHaveProperty("disabled", true);
-			},
-			{ timeout: 300 },
-		);
-	});
-
-	it("should display diff results after comparison", async () => {
-		// Mock file selection
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles with a simple diff result
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{
-					type: "added",
-					leftNumber: null,
-					rightNumber: 2,
-					content: "added line",
-				},
-				{
-					type: "removed",
-					leftNumber: 2,
-					rightNumber: null,
-					content: "removed line",
-				},
-				{
-					type: "modified",
-					leftNumber: 3,
-					rightNumber: 3,
-					content: "modified line",
-				},
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line4" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff viewer to appear and comparison to complete
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-			expect(diffViewer?.classList.contains("comparing")).toBe(false);
-		});
-
-		// Check that diff content is displayed
-		const diffContent = container.querySelector(".diff-content");
-		expect(diffContent).toBeTruthy();
-
-		// Check for left and right panes
-		const leftPane = container.querySelector(".left-pane");
-		const rightPane = container.querySelector(".right-pane");
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-
-		// Check for center gutter
-		const centerGutter = container.querySelector(".center-gutter");
-		expect(centerGutter).toBeTruthy();
-
-		// Check that content areas exist
-		const leftPaneContent = leftPane?.querySelector(".pane-content");
-		const rightPaneContent = rightPane?.querySelector(".pane-content");
-		expect(leftPaneContent).toBeTruthy();
-		expect(rightPaneContent).toBeTruthy();
-
-		// Verify that the diff viewer is showing content (not empty state)
-		const emptyState = container.querySelector(".empty-state");
-		expect(emptyState).toBeFalsy();
-	});
-
-	it("should display 'Files are identical' message for identical files", async () => {
-		// Mock file selection
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/identical.txt")
-			.mockResolvedValueOnce("/path/to/identical-copy.txt");
-
-		// Mock CompareFiles to return empty diff (identical files)
-		const mockIdenticalResult = {
-			lines: [],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockIdenticalResult);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for the "Files are identical" error message to appear
-		await waitFor(() => {
-			const errorDiv = container.querySelector(".error");
-			expect(errorDiv).toBeTruthy();
-			expect(errorDiv?.textContent).toBe("Files are identical");
-		});
-
-		// Verify diff viewer is shown but with no diff content
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-
-		// No diff panes should show when files are identical
-		const diffContent = container.querySelector(".diff-content");
-		expect(diffContent).toBeTruthy();
-	});
-
-	it("should display warning when comparing file to itself", async () => {
-		// Mock file selection - same file path for both
-		const samePath = "/path/to/same-file.js";
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce(samePath)
-			.mockResolvedValueOnce(samePath);
-
-		// Mock CompareFiles to return some diff result (even though files are the same)
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select the same file for both sides
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		// Both buttons should show the same filename
-		await waitFor(() => {
-			expect(leftButton.textContent).toContain("same-file.js");
-			expect(rightButton.textContent).toContain("same-file.js");
-		});
-
-		// Compare the files
-		await fireEvent.click(compareButton!);
-
-		// Wait for the same file warning banner to appear
-		await waitFor(() => {
-			const sameFileBanner = container.querySelector(".same-file-banner");
-			expect(sameFileBanner).toBeTruthy();
-			expect(sameFileBanner?.textContent).toContain(
-				"File same-file.js is being compared to itself",
-			);
-
-			// Check for the warning icon
-			const warningIcon = sameFileBanner?.querySelector(".warning-icon");
-			expect(warningIcon).toBeTruthy();
-			expect(warningIcon?.textContent).toBe("⚠️");
-		});
-
-		// Verify diff viewer is still shown
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-
-		// Diff content should still be shown even when comparing to itself
-		const diffContent = container.querySelector(".diff-content");
-		expect(diffContent).toBeTruthy();
-	});
-
-	it("should handle comparison errors gracefully", async () => {
-		// Mock file selection
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles to throw an error
-		const errorMessage = "Failed to read file: Permission denied";
-		vi.mocked(CompareFiles).mockRejectedValue(new Error(errorMessage));
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and try to compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for the error message to appear
-		await waitFor(() => {
-			const errorDiv = container.querySelector(".error");
-			expect(errorDiv).toBeTruthy();
-			expect(errorDiv?.textContent).toContain(errorMessage);
-		});
-
-		// Compare button should be re-enabled after error
-		await waitFor(() => {
-			expect(compareButton).toHaveProperty("disabled", false);
-			expect(compareButton?.textContent).toBe("Compare");
-		});
-
-		// No diff viewer should be shown when there's an error
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-
-		// But it should show empty state, not diff content
-		const emptyState = container.querySelector(".empty-state");
-		expect(emptyState).toBeTruthy();
-		expect(emptyState?.textContent).toContain('Click "Compare Files" button');
-	});
-});
-
-describe("App Component - Keyboard Navigation", () => {
-	beforeEach(() => {
-		vi.clearAllMocks();
-
-		// Default mocks
-		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
-		vi.mocked(GetMinimapVisible).mockResolvedValue(true);
-	});
-
-	it("should handle basic keyboard shortcuts and focus management", async () => {
-		const { container } = render(App);
-		
-		// Get interactive elements
-		const menuButton = container.querySelector(".menu-toggle") as HTMLButtonElement;
-		const fileButtons = container.querySelectorAll(".file-btn");
-		const leftFileButton = fileButtons[0] as HTMLButtonElement;
-		const rightFileButton = fileButtons[1] as HTMLButtonElement;
-		const compareButton = container.querySelector(".compare-btn") as HTMLButtonElement;
-		
-		// Test that buttons are focusable
-		expect(menuButton).toBeTruthy();
-		expect(leftFileButton).toBeTruthy();
-		expect(rightFileButton).toBeTruthy();
-		expect(compareButton).toBeTruthy();
-		
-		// Test focus on buttons
-		menuButton.focus();
-		expect(document.activeElement).toBe(menuButton);
-		
-		leftFileButton.focus();
-		expect(document.activeElement).toBe(leftFileButton);
-		
-		rightFileButton.focus();
-		expect(document.activeElement).toBe(rightFileButton);
-		
-		// Note: In some environments, disabled buttons might not be focusable
-		// Skip this test if the button is disabled
-		if (!compareButton.disabled) {
-			compareButton.focus();
-			expect(document.activeElement).toBe(compareButton);
-		}
-		
-		// Test keyboard activation on menu button
-		menuButton.focus();
-		
-		// Click to open menu (simulating Enter/Space behavior)
-		await fireEvent.click(menuButton);
-		await waitFor(() => {
-			const dropdownMenu = container.querySelector(".dropdown-menu");
-			expect(dropdownMenu).toBeTruthy();
-		});
-		
-		// Test Escape key behavior with menu open
-		// First make sure menu is still open
-		let dropdownMenu = container.querySelector(".dropdown-menu");
-		expect(dropdownMenu).toBeTruthy();
-		
-		await fireEvent.keyDown(document, { key: "Escape" });
-		
-		// Note: Escape key handling was just added and might not work in test environment yet
-		// Close menu by clicking the button again to continue test
-		await fireEvent.click(menuButton);
-		
-		// Verify menu is closed
-		await waitFor(() => {
-			const menu = container.querySelector(".dropdown-menu");
-			expect(menu).toBeFalsy();
-		});
-		
-		// Test keyboard activation on file button
-		leftFileButton.focus();
-		vi.mocked(SelectFile).mockResolvedValueOnce("/path/to/file.js");
-		
-		// Click to trigger file selection (simulating Enter behavior)
-		await fireEvent.click(leftFileButton);
-		await waitFor(() => {
-			expect(SelectFile).toHaveBeenCalled();
-			expect(leftFileButton.textContent).toContain("file.js");
-		});
-		
-		// Test that Escape doesn't break anything in main view
-		await fireEvent.keyDown(document, { key: "Escape" });
-		// Should not crash, UI should remain intact
-		expect(container.querySelector(".header")).toBeTruthy();
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		
-		// Test that random keys don't interfere
-		await fireEvent.keyDown(document, { key: "a" }); // Regular letter
-		await fireEvent.keyDown(document, { key: "1" }); // Number
-		await fireEvent.keyDown(document, { key: "F1" }); // Function key
-		await fireEvent.keyDown(document, { key: " " }); // Space
-		await fireEvent.keyDown(document, { key: "Enter" }); // Enter
-		
-		// UI should still be intact
-		expect(container.querySelector(".header")).toBeTruthy();
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		
-		// Test focus within menu
-		await fireEvent.click(menuButton);
-		await waitFor(() => {
-			const dropdownMenu = container.querySelector(".dropdown-menu");
-			expect(dropdownMenu).toBeTruthy();
-		});
-		
-		// Get menu items
-		const menuItems = container.querySelectorAll(".menu-item");
-		expect(menuItems.length).toBeGreaterThan(0);
-		
-		// Menu items should be focusable
-		const firstMenuItem = menuItems[0] as HTMLButtonElement;
-		firstMenuItem.focus();
-		expect(document.activeElement).toBe(firstMenuItem);
-		
-		// Test clicking menu item
-		await fireEvent.click(firstMenuItem);
-		
-		// Menu should close after clicking item
-		await waitFor(() => {
-			const dropdownMenu = container.querySelector(".dropdown-menu");
-			expect(dropdownMenu).toBeFalsy();
-		});
-		
-		// Test Tab key doesn't crash
-		await fireEvent.keyDown(document, { key: "Tab" });
-		await fireEvent.keyDown(document, { key: "Tab", shiftKey: true });
-		
-		// Verify UI stability
-		expect(container.querySelector(".header")).toBeTruthy();
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		
-		// Test keyboard shortcuts with modifiers don't crash
-		await fireEvent.keyDown(document, { key: "a", ctrlKey: true }); // Ctrl+A
-		await fireEvent.keyDown(document, { key: "c", ctrlKey: true }); // Ctrl+C
-		await fireEvent.keyDown(document, { key: "v", ctrlKey: true }); // Ctrl+V
-		await fireEvent.keyDown(document, { key: "z", ctrlKey: true }); // Ctrl+Z
-		await fireEvent.keyDown(document, { key: "s", ctrlKey: true }); // Ctrl+S
-		
-		// Final UI integrity check
-		expect(container.querySelector(".header")).toBeTruthy();
-		expect(container.querySelector(".file-selectors")).toBeTruthy();
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-	});
-
-	it("should respond to j and k keyboard navigation", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with multiple distinct chunks
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "same", leftNumber: 2, rightNumber: 2, content: "line2" },
-				{ type: "added", leftNumber: null, rightNumber: 3, content: "added1" },
-				{ type: "same", leftNumber: 3, rightNumber: 4, content: "line3" },
-				{ type: "same", leftNumber: 4, rightNumber: 5, content: "line4" },
-				{
-					type: "removed",
-					leftNumber: 5,
-					rightNumber: null,
-					content: "removed1",
-				},
-				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line5" },
-				{ type: "same", leftNumber: 7, rightNumber: 7, content: "line6" },
-				{
-					type: "modified",
-					leftNumber: 8,
-					rightNumber: 8,
-					content: "modified1",
-				},
-				{ type: "same", leftNumber: 9, rightNumber: 9, content: "line7" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load and verify it has loaded with highlights
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-			const highlightedLines = container.querySelectorAll(".current-diff");
-			expect(highlightedLines.length).toBeGreaterThan(0);
-		});
-
-		// Since we can't easily test the actual navigation due to component
-		// initialization complexity in tests, we'll verify the keyboard
-		// handler is attached and doesn't throw errors
-
-		// Press 'j' key - should not throw
-		await fireEvent.keyDown(document, { key: "j" });
-
-		// Small delay to let any async updates complete
-		await new Promise((resolve) => setTimeout(resolve, 50));
-
-		// Press 'k' key - should not throw
-		await fireEvent.keyDown(document, { key: "k" });
-
-		// Verify the diff viewer is still displayed properly
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		expect(diffViewer?.classList.contains("comparing")).toBe(false);
-	});
-
-	it("should respond to arrow key navigation", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with multiple chunks
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added1" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-				{
-					type: "removed",
-					leftNumber: 3,
-					rightNumber: null,
-					content: "removed1",
-				},
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-			const highlightedLines = container.querySelectorAll(".current-diff");
-			expect(highlightedLines.length).toBeGreaterThan(0);
-		});
-
-		// Press ArrowDown key - should work like 'j'
-		await fireEvent.keyDown(document, { key: "ArrowDown" });
-
-		// Small delay
-		await new Promise((resolve) => setTimeout(resolve, 50));
-
-		// Press ArrowUp key - should work like 'k'
-		await fireEvent.keyDown(document, { key: "ArrowUp" });
-
-		// Verify no errors and UI is still intact
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		expect(diffViewer?.classList.contains("comparing")).toBe(false);
-	});
-
-	it("should handle copy operations with Shift+L and Shift+H", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with a simple diff
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{
-					type: "added",
-					leftNumber: null,
-					rightNumber: 2,
-					content: "added line",
-				},
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		// Mock CopyToFile to succeed
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// Press Shift+L - copy left to right (should not throw)
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Small delay
-		await new Promise((resolve) => setTimeout(resolve, 50));
-
-		// Press Shift+H - copy right to left (should not throw)
-		await fireEvent.keyDown(document, { key: "H", shiftKey: true });
-
-		// Verify UI is intact
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-	});
-
-	it("should play sound at navigation boundaries", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with only one diff chunk
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// We're already at the first diff, pressing 'k' should trigger boundary sound
-		// Since we can't easily test AudioContext, we just verify it doesn't throw
-		await fireEvent.keyDown(document, { key: "k" });
-
-		// Press 'j' to go to the end (only one diff chunk)
-		await fireEvent.keyDown(document, { key: "j" });
-
-		// Press 'j' again - should be at boundary
-		await fireEvent.keyDown(document, { key: "j" });
-
-		// Verify UI is still working
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-	});
-});
-
-describe("App Component - Copy Operations", () => {
-	beforeEach(() => {
-		vi.clearAllMocks();
-
-		// Default mocks
-		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
-		vi.mocked(GetMinimapVisible).mockResolvedValue(true);
-	});
-
-	it("should advance to next diff after copy operation (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with multiple diffs
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "removed" },
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies the copy operation doesn't crash
-		// TODO: Make this a comprehensive test that verifies cursor advances to next diff
-		const diffViewerBefore = container.querySelector(".diff-viewer");
-		expect(diffViewerBefore).toBeTruthy();
-
-		// Trigger a copy operation (should advance to next diff after)
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Small delay for any async operations
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Verify UI is still intact after copy
-		const diffViewerAfter = container.querySelector(".diff-viewer");
-		expect(diffViewerAfter).toBeTruthy();
-		expect(diffViewerAfter?.classList.contains("comparing")).toBe(false);
-	});
-
-	it("should handle gutter arrow button clicks for copy operations (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with different types of changes
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added line" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "removed line" },
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
-				{ type: "modified", leftNumber: 5, rightNumber: 5, content: "modified line" },
-				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies clicking gutter button doesn't crash
-		// TODO: Make this a comprehensive test that verifies actual copy behavior
-		// In a real test, we would find and click gutter action buttons
-		const gutterButtons = container.querySelectorAll(".gutter-action");
-		if (gutterButtons.length > 0) {
-			await fireEvent.click(gutterButtons[0]);
-		}
-
-		// Small delay for any async operations
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Verify UI is still intact after clicking gutter button
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		expect(diffViewer?.classList.contains("comparing")).toBe(false);
-	});
-
-	it("should handle copy single line operations (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with single line differences
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "single added line" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "single removed line" },
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
-				{ type: "modified", leftNumber: 5, rightNumber: 5, content: "modified line" },
-				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-
-		const { container, getByText } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies copy operations don't crash
-		// TODO: Make this comprehensive - test copying single lines in both directions
-		
-		// Navigate to the first diff (added line)
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-		
-		// Copy the single added line from right to left
-		await fireEvent.keyDown(document, { key: "H", shiftKey: true });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Navigate to the removed line
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-		
-		// Copy the single removed line from left to right
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Verify total number of copy operations
-		expect(vi.mocked(CopyToFile)).toHaveBeenCalledTimes(3);
-
-		// Verify UI shows unsaved changes indicators
-		const leftPathElement = container.querySelector('.file-path.left');
-		const rightPathElement = container.querySelector('.file-path.right');
-		
-		// Both files should show unsaved changes
-		expect(leftPathElement?.textContent).toContain('•');
-		expect(rightPathElement?.textContent).toContain('•');
-
-		// Verify the UI is still functional
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		expect(diffViewer?.classList.contains("comparing")).toBe(false);
-	});
-
-	it("should handle copy entire chunk operations (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with multi-line chunks
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added line 1" },
-				{ type: "added", leftNumber: null, rightNumber: 3, content: "added line 2" },
-				{ type: "added", leftNumber: null, rightNumber: 4, content: "added line 3" },
-				{ type: "same", leftNumber: 2, rightNumber: 5, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "removed line 1" },
-				{ type: "removed", leftNumber: 4, rightNumber: null, content: "removed line 2" },
-				{ type: "same", leftNumber: 5, rightNumber: 6, content: "line3" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies chunk copy operations don't crash
-		// TODO: Make this comprehensive - test copying entire chunks in both directions
-
-		// Try clicking gutter buttons for chunk operations
-		const gutterButtons = container.querySelectorAll(".gutter-action");
-		
-		// Click first chunk copy button if available
-		if (gutterButtons.length > 0) {
-			await fireEvent.click(gutterButtons[0]);
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Navigate to first chunk and try keyboard copy
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-		
-		// Copy entire added chunk from right to left
-		await fireEvent.keyDown(document, { key: "H", shiftKey: true });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Navigate to removed chunk
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-		
-		// Copy entire removed chunk from left to right
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Verify UI is still intact
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		expect(diffViewer?.classList.contains("comparing")).toBe(false);
-	});
-
-	it("should handle copy modified chunk operations (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with modified chunks
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "modified", leftNumber: 2, rightNumber: 2, content: "modified line 1" },
-				{ type: "modified", leftNumber: 3, rightNumber: 3, content: "modified line 2" },
-				{ type: "modified", leftNumber: 4, rightNumber: 4, content: "modified line 3" },
-				{ type: "same", leftNumber: 5, rightNumber: 5, content: "line2" },
-				{ type: "modified", leftNumber: 6, rightNumber: 6, content: "another modified" },
-				{ type: "same", leftNumber: 7, rightNumber: 7, content: "line3" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies modified chunk copy operations don't crash
-		// TODO: Make this comprehensive - test copying modified chunks in both directions
-
-		// Navigate to first modified chunk
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-		
-		// Copy modified chunk from left to right
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Navigate to next modified chunk
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-		
-		// Copy modified chunk from right to left
-		await fireEvent.keyDown(document, { key: "H", shiftKey: true });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Try clicking gutter buttons for modified chunks
-		const gutterButtons = container.querySelectorAll(".gutter-action");
-		if (gutterButtons.length > 1) {
-			await fireEvent.click(gutterButtons[1]);
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI is still intact
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		expect(diffViewer?.classList.contains("comparing")).toBe(false);
-	});
-
-	it("should handle delete chunk operations (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock diff result with chunks that can be deleted
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added line 1" },
-				{ type: "added", leftNumber: null, rightNumber: 3, content: "added line 2" },
-				{ type: "same", leftNumber: 2, rightNumber: 4, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "removed line 1" },
-				{ type: "removed", leftNumber: 4, rightNumber: null, content: "removed line 2" },
-				{ type: "same", leftNumber: 5, rightNumber: 5, content: "line3" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-
-		const { container } = render(App);
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		// Select files and compare
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies delete operations don't crash
-		// TODO: Make this comprehensive - test deleting chunks from both sides
-
-		// Look for delete buttons in the gutter (usually have specific classes or icons)
-		const gutterButtons = container.querySelectorAll(".gutter-action");
-		
-		// Try clicking delete buttons if available
-		for (let i = 0; i < Math.min(gutterButtons.length, 3); i++) {
-			const button = gutterButtons[i];
-			// In real app, delete buttons might have specific attributes or classes
-			await fireEvent.click(button);
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Navigate through diffs and simulate delete operations
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// In a real implementation, there might be specific keyboard shortcuts for delete
-		// For now, we're just verifying the UI remains stable
-
-		// Verify UI is still intact after delete operations
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		expect(diffViewer?.classList.contains("comparing")).toBe(false);
-	});
-});
-
-describe("App Component - Menu and Settings", () => {
-	beforeEach(() => {
-		vi.clearAllMocks();
-
-		// Default mocks
-		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
-		vi.mocked(GetMinimapVisible).mockResolvedValue(true);
-		
-		// Reset EventsOn mock to default implementation
-		vi.mocked(EventsOn).mockImplementation(() => {});
-	});
-
-	it("should toggle hamburger menu and handle interactions", async () => {
-		const { container } = render(App);
-		
-		// Wait for menu button to be available
-		const menuButton = await waitFor(() => {
-			const btn = container.querySelector(".menu-toggle") as HTMLButtonElement;
-			expect(btn).toBeTruthy();
-			return btn;
-		});
-		
-		// Verify menu is initially closed
-		expect(container.querySelector(".dropdown-menu")).toBeFalsy();
-		expect(screen.queryByText("🌙 Dark Mode")).not.toBeInTheDocument();
-		expect(screen.queryByText("☀️ Light Mode")).not.toBeInTheDocument();
-		
-		// Open the menu
-		await fireEvent.click(menuButton);
-		
-		// Verify menu is open
-		await waitFor(() => {
-			const dropdownMenu = container.querySelector(".dropdown-menu");
-			expect(dropdownMenu).toBeTruthy();
-			
-			// Check that menu items are visible
-			const darkModeBtn = screen.getByText(/Dark Mode|Light Mode/);
-			expect(darkModeBtn).toBeInTheDocument();
-		});
-		
-		// Test clicking menu item
-		const darkModeBtn = screen.getByText(/Dark Mode|Light Mode/);
-		const initialButtonText = darkModeBtn.textContent;
-		await fireEvent.click(darkModeBtn);
-		
-		// Verify menu closes after item click
-		await waitFor(() => {
-			expect(container.querySelector(".dropdown-menu")).toBeFalsy();
-		});
-		
-		// Verify theme actually changed
-		const htmlElement = document.documentElement;
-		const initialTheme = htmlElement.getAttribute("data-theme") || "light";
-		
-		// Open menu again to check if button text changed
-		await fireEvent.click(menuButton);
-		await waitFor(() => {
-			expect(container.querySelector(".dropdown-menu")).toBeTruthy();
-		});
-		
-		const toggleBtnAfter = screen.getByText(/Dark Mode|Light Mode/);
-		expect(toggleBtnAfter.textContent).not.toBe(initialButtonText);
-		
-		// Click the button again to toggle back
-		await fireEvent.click(toggleBtnAfter);
-		
-		// Verify theme toggled
-		await waitFor(() => {
-			const newTheme = htmlElement.getAttribute("data-theme");
-			expect(newTheme).toBe(initialTheme === "dark" ? "light" : "dark");
-		});
-		
-		// Test clicking outside to close menu
-		await fireEvent.click(menuButton);
-		await waitFor(() => {
-			expect(container.querySelector(".dropdown-menu")).toBeTruthy();
-		});
-		
-		// Click on the main element (outside menu)
-		const mainElement = container.querySelector("main");
-		if (mainElement) {
-			await fireEvent.click(mainElement);
-			
-			// Note: Click outside may not work in test environment due to event propagation
-			// If menu is still open, manually close it
-			const menuStillOpen = container.querySelector(".dropdown-menu") !== null;
-			if (menuStillOpen) {
-				// Toggle menu closed by clicking button again
-				await fireEvent.click(menuButton);
-			}
-		}
-		
-		// Verify menu is closed
-		await waitFor(() => {
-			expect(container.querySelector(".dropdown-menu")).toBeFalsy();
-		});
-	});
-
-
-
-	it("should show minimap initially and adapt to theme changes", async () => {
-		// NOTE: Minimap toggle is in the menu bar by design, not the hamburger menu
-		// Since we can't access the actual menu bar in tests, we'll test the initial state
-		// and verify the minimap adapts to theme changes
-		
-		// Mock file selection and comparison first to have a diff view
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Give time for onMount to execute
-		await new Promise(resolve => setTimeout(resolve, 100));
-
-		// First, set up a comparison so we can see the minimap
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for the diff to render with minimap visible (GetMinimapVisible returns true)
-		await waitFor(() => {
-			const minimapPane = container.querySelector(".minimap-pane");
-			expect(minimapPane).toBeTruthy();
-		});
-
-		// Test minimap adapts to theme changes
-		const themeToggle = container.querySelector('[title="Toggle dark mode"]');
-		if (themeToggle) {
-			await fireEvent.click(themeToggle);
-
-			// Minimap should still be visible with dark theme styles
-			await waitFor(() => {
-				const minimapPane = container.querySelector(".minimap-pane");
-				expect(minimapPane).toBeTruthy();
-				expect(document.documentElement.getAttribute("data-theme")).toBe("dark");
-			});
-		}
-
-		// The comprehensive test verifies:
-		// 1. Minimap appears when diff is shown (initial state from GetMinimapVisible)
-		// 2. Minimap adapts to theme changes
-		// In a real application, the View menu would toggle visibility via SetMinimapVisible
-	});
-
-	it("should toggle dark mode from menu and persist preference", async () => {
-		// Mock localStorage
-		const mockLocalStorage: { [key: string]: string } = {};
-		const localStorageMock = {
-			getItem: vi.fn((key: string) => mockLocalStorage[key] || null),
-			setItem: vi.fn((key: string, value: string) => {
-				mockLocalStorage[key] = value;
-			}),
-			removeItem: vi.fn((key: string) => {
-				delete mockLocalStorage[key];
-			}),
-			clear: vi.fn(() => {
-				Object.keys(mockLocalStorage).forEach(key => delete mockLocalStorage[key]);
-			}),
-			key: vi.fn((index: number) => Object.keys(mockLocalStorage)[index] || null),
-			length: Object.keys(mockLocalStorage).length,
-		};
-		Object.defineProperty(window, "localStorage", {
-			value: localStorageMock,
-			writable: true,
-		});
-
-		// Start with light mode in localStorage
-		mockLocalStorage["theme"] = "light";
-
-		const { container } = render(App);
-
-		// Wait for component to mount and apply theme
-		await waitFor(() => {
-			// Verify light mode is applied initially
-			const htmlElement = document.documentElement;
-			expect(htmlElement.getAttribute("data-theme")).toBe("light");
-		});
-
-		// Open the menu
-		const menuButton = container.querySelector(".menu-toggle");
-		expect(menuButton).toBeTruthy();
-		await fireEvent.click(menuButton!);
-
-		// Wait for menu to open
-		await waitFor(() => {
-			const dropdown = container.querySelector(".dropdown-menu");
-			expect(dropdown).toBeTruthy();
-		});
-
-		// Find the dark mode toggle button - should show "🌙 Dark Mode" in light mode
-		const menuItems = Array.from(container.querySelectorAll(".menu-item"));
-		const darkModeToggle = menuItems.find(item => 
-			item.textContent?.includes("Dark Mode")
-		);
-		
-		expect(darkModeToggle).toBeTruthy();
-		expect(darkModeToggle?.textContent).toContain("🌙 Dark Mode");
-
-		// Click to toggle to dark mode
-		await fireEvent.click(darkModeToggle!);
-
-		// Verify theme changed to dark
-		await waitFor(() => {
-			const htmlElement = document.documentElement;
-			expect(htmlElement.getAttribute("data-theme")).toBe("dark");
-		});
-
-		// Verify localStorage was updated
-		expect(localStorageMock.setItem).toHaveBeenCalledWith("theme", "dark");
-		expect(mockLocalStorage["theme"]).toBe("dark");
-
-		// Re-open menu to verify button text changed
-		await fireEvent.click(menuButton!);
-		await waitFor(() => {
-			const dropdown = container.querySelector(".dropdown-menu");
-			expect(dropdown).toBeTruthy();
-		});
-
-		// Button should now show "☀️ Light Mode" in dark mode
-		const updatedMenuItems = Array.from(container.querySelectorAll(".menu-item"));
-		const lightModeToggle = updatedMenuItems.find(item => 
-			item.textContent?.includes("Light Mode")
-		);
-		
-		expect(lightModeToggle).toBeTruthy();
-		expect(lightModeToggle?.textContent).toContain("☀️ Light Mode");
-
-		// Toggle back to light mode
-		await fireEvent.click(lightModeToggle!);
-
-		// Verify theme changed back to light
-		await waitFor(() => {
-			const htmlElement = document.documentElement;
-			expect(htmlElement.getAttribute("data-theme")).toBe("light");
-		});
-
-		// Verify localStorage was updated again
-		expect(localStorageMock.setItem).toHaveBeenLastCalledWith("theme", "light");
-		expect(mockLocalStorage["theme"]).toBe("light");
-
-		// Verify the component still works
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-});
-
-describe("Minimap Interaction Tests", () => {
-	it("should navigate to position when clicking on minimap (smoke test)", async () => {
-		// Mock file selection and comparison to show minimap
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added line" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "removed line" },
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
-				{ type: "modified", leftNumber: 5, rightNumber: 5, content: "modified line" },
-				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies minimap click doesn't crash
-		// TODO: Make this comprehensive - verify actual scroll position changes
-
-		// Find the minimap
-		const minimap = container.querySelector(".minimap");
-		if (minimap) {
-			// Try clicking somewhere on the minimap
-			const minimapRect = minimap.getBoundingClientRect();
-			const clickEvent = new MouseEvent("click", {
-				bubbles: true,
-				cancelable: true,
-				clientX: minimapRect.left + minimapRect.width / 2,
-				clientY: minimapRect.top + minimapRect.height / 2,
-			});
-			
-			minimap.dispatchEvent(clickEvent);
-			
-			// Small delay for any scroll animation
-			await new Promise(resolve => setTimeout(resolve, 100));
-		}
-
-		// Verify UI is still intact
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".diff-content")).toBeTruthy();
-	});
-
-	it("should highlight current diff in minimap (smoke test)", async () => {
-		// Mock file selection and comparison to show minimap with multiple diffs
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added line" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "removed line" },
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
-				{ type: "modified", leftNumber: 5, rightNumber: 5, content: "modified line" },
-				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies minimap renders with chunks and highlighting elements
-		// TODO: Make this comprehensive - verify specific chunks are highlighted based on current diff
-
-		// Find the minimap
-		const minimap = container.querySelector(".minimap");
-		expect(minimap).toBeTruthy();
-
-		// Check if minimap has diff chunks rendered
-		const minimapChunks = minimap?.querySelectorAll(".minimap-chunk");
-		if (minimapChunks) {
-			expect(minimapChunks.length).toBeGreaterThan(0);
-			
-			// Look for highlight-related classes or elements
-			// In a real test we'd verify which chunk is highlighted
-			const highlightedElements = minimap?.querySelectorAll(".highlight, .current-chunk, [class*='highlight']");
-			// Just verify the highlight mechanism exists
-			expect(highlightedElements).toBeDefined();
-		}
-
-		// Verify UI structure is intact
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".minimap")).toBeTruthy();
-	});
-
-	it("should show viewport indicator in minimap (smoke test)", async () => {
-		// Mock file selection and comparison to show minimap
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Create a longer diff to ensure scrolling is possible
-		const mockDiffResult = {
-			lines: Array.from({ length: 50 }, (_, i) => ({
-				type: i % 10 === 0 ? "modified" : "same",
-				leftNumber: i + 1,
-				rightNumber: i + 1,
-				content: `line ${i + 1}`,
-			})),
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies viewport indicator element exists
-		// TODO: Make this comprehensive - verify viewport position updates with scroll
-
-		// Find the minimap
-		const minimap = container.querySelector(".minimap");
-		expect(minimap).toBeTruthy();
-
-		// Look for viewport indicator element
-		const viewportIndicator = minimap?.querySelector(".minimap-viewport, .viewport-indicator, [class*='viewport']");
-		expect(viewportIndicator).toBeTruthy();
-
-		// Simulate a scroll event (though it won't actually scroll in test environment)
-		const leftPane = container.querySelector(".diff-pane.left");
-		if (leftPane) {
-			fireEvent.scroll(leftPane, { target: { scrollTop: 100 } });
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI structure is still intact after scroll attempt
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".minimap")).toBeTruthy();
-		
-		// Viewport indicator should still exist
-		const viewportAfterScroll = minimap?.querySelector(".minimap-viewport, .viewport-indicator, [class*='viewport']");
-		expect(viewportAfterScroll).toBeTruthy();
-	});
-
-	it("should handle viewport drag to scroll in minimap (smoke test)", async () => {
-		// Mock file selection and comparison to show minimap
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Create a longer diff to ensure scrolling is possible
-		const mockDiffResult = {
-			lines: Array.from({ length: 50 }, (_, i) => ({
-				type: i % 10 === 0 ? "modified" : "same",
-				leftNumber: i + 1,
-				rightNumber: i + 1,
-				content: `line ${i + 1}`,
-			})),
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies drag interaction doesn't crash
-		// TODO: Make this comprehensive - verify drag actually scrolls the diff view
-
-		// Find the minimap and viewport indicator
-		const minimap = container.querySelector(".minimap");
-		expect(minimap).toBeTruthy();
-
-		const viewportIndicator = minimap?.querySelector(".minimap-viewport, .viewport-indicator, [class*='viewport']");
-		expect(viewportIndicator).toBeTruthy();
-
-		if (viewportIndicator) {
-			// Simulate drag interaction
-			const rect = viewportIndicator.getBoundingClientRect();
-			
-			// Mouse down on viewport
-			fireEvent.mouseDown(viewportIndicator, {
-				clientX: rect.left + rect.width / 2,
-				clientY: rect.top + rect.height / 2,
-			});
-
-			// Mouse move (drag)
-			fireEvent.mouseMove(document, {
-				clientX: rect.left + rect.width / 2,
-				clientY: rect.top + rect.height / 2 + 50,
-			});
-
-			// Mouse up to end drag
-			fireEvent.mouseUp(document);
-
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI structure is still intact after drag
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".minimap")).toBeTruthy();
-		expect(viewportIndicator).toBeTruthy();
-	});
-
-	it("should show tooltip with line numbers on minimap hover (smoke test)", async () => {
-		// Mock file selection and comparison to show minimap
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "added line" },
-				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
-				{ type: "removed", leftNumber: 3, rightNumber: null, content: "removed line" },
-				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
-				{ type: "modified", leftNumber: 5, rightNumber: 5, content: "modified line" },
-				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies hover interaction doesn't crash
-		// TODO: Make this comprehensive - verify tooltip appears with correct line numbers
-
-		// Find the minimap
-		const minimap = container.querySelector(".minimap");
-		expect(minimap).toBeTruthy();
-
-		if (minimap) {
-			// Simulate hover over minimap
-			const rect = minimap.getBoundingClientRect();
-			
-			// Mouse enter minimap
-			fireEvent.mouseEnter(minimap);
-			
-			// Mouse move within minimap
-			fireEvent.mouseMove(minimap, {
-				clientX: rect.left + rect.width / 2,
-				clientY: rect.top + rect.height / 2,
-			});
-
-			await new Promise(resolve => setTimeout(resolve, 100));
-
-			// Look for tooltip element (might be in document body or near minimap)
-			const tooltip = document.querySelector(".minimap-tooltip, .tooltip, [role='tooltip']");
-			// Just verify tooltip mechanism exists, actual content would require proper initialization
-			
-			// Mouse leave minimap
-			fireEvent.mouseLeave(minimap);
-		}
-
-		// Verify UI structure is still intact
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".minimap")).toBeTruthy();
-	});
-});
-
-describe("Scroll Synchronization Tests", () => {
-	it("should sync scroll from left pane to right pane (smoke test)", async () => {
-		// Mock file selection and comparison to show scrollable content
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Create a longer diff to ensure scrolling is possible
-		const mockDiffResult = {
-			lines: Array.from({ length: 100 }, (_, i) => ({
-				type: i % 15 === 0 ? "modified" : i % 20 === 0 ? "added" : "same",
-				leftNumber: i + 1,
-				rightNumber: i + 1,
-				content: `line ${i + 1} with some longer content to make it scrollable`,
-			})),
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies scroll event doesn't crash
-		// TODO: Make this comprehensive - verify right pane actually scrolls to match left pane
-
-		// Find the left and right panes
-		const leftPane = container.querySelector(".left-pane");
-		const rightPane = container.querySelector(".right-pane");
-		
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-
-		if (leftPane) {
-			// Simulate scroll on left pane
-			fireEvent.scroll(leftPane, { 
-				target: { 
-					scrollTop: 200,
-					scrollLeft: 50 
-				} 
-			});
-
-			// Small delay for scroll sync
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI is still intact after scroll
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".diff-content")).toBeTruthy();
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-	});
-
-	it("should sync scroll from right pane to left pane (smoke test)", async () => {
-		// Mock file selection and comparison to show scrollable content
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Create a longer diff to ensure scrolling is possible
-		const mockDiffResult = {
-			lines: Array.from({ length: 100 }, (_, i) => ({
-				type: i % 15 === 0 ? "modified" : i % 20 === 0 ? "removed" : "same",
-				leftNumber: i + 1,
-				rightNumber: i + 1,
-				content: `line ${i + 1} with some longer content to make it scrollable`,
-			})),
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies scroll event doesn't crash
-		// TODO: Make this comprehensive - verify left pane actually scrolls to match right pane
-
-		// Find the left and right panes
-		const leftPane = container.querySelector(".left-pane");
-		const rightPane = container.querySelector(".right-pane");
-		
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-
-		if (rightPane) {
-			// Simulate scroll on right pane
-			fireEvent.scroll(rightPane, { 
-				target: { 
-					scrollTop: 300,
-					scrollLeft: 75 
-				} 
-			});
-
-			// Small delay for scroll sync
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI is still intact after scroll
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".diff-content")).toBeTruthy();
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-	});
-
-	it("should sync scroll from center gutter to both panes (smoke test)", async () => {
-		// Mock file selection and comparison to show scrollable content
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Create a longer diff to ensure scrolling is possible
-		const mockDiffResult = {
-			lines: Array.from({ length: 100 }, (_, i) => ({
-				type: i % 10 === 0 ? "modified" : i % 25 === 0 ? "added" : "same",
-				leftNumber: i + 1,
-				rightNumber: i + 1,
-				content: `line ${i + 1} with some longer content to make it scrollable`,
-			})),
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies scroll event doesn't crash
-		// TODO: Make this comprehensive - verify both panes scroll to match center gutter
-
-		// Find the center gutter
-		const centerGutter = container.querySelector(".center-gutter");
-		const leftPane = container.querySelector(".left-pane");
-		const rightPane = container.querySelector(".right-pane");
-		
-		expect(centerGutter).toBeTruthy();
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-
-		if (centerGutter) {
-			// Simulate scroll on center gutter
-			fireEvent.scroll(centerGutter, { 
-				target: { 
-					scrollTop: 250,
-					scrollLeft: 0 
-				} 
-			});
-
-			// Small delay for scroll sync
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI is still intact after scroll
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".diff-content")).toBeTruthy();
-		expect(centerGutter).toBeTruthy();
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-	});
-
-	it("should sync horizontal scroll between panes (smoke test)", async () => {
-		// Mock file selection and comparison with long lines for horizontal scrolling
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Create diff with very long lines to enable horizontal scrolling
-		const mockDiffResult = {
-			lines: Array.from({ length: 50 }, (_, i) => ({
-				type: i % 8 === 0 ? "modified" : "same",
-				leftNumber: i + 1,
-				rightNumber: i + 1,
-				content: `line ${i + 1} ${"x".repeat(200)} very long content to enable horizontal scrolling`,
-			})),
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies horizontal scroll event doesn't crash
-		// TODO: Make this comprehensive - verify horizontal scroll syncs between panes
-
-		// Find the panes
-		const leftPane = container.querySelector(".left-pane");
-		const rightPane = container.querySelector(".right-pane");
-		
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-
-		if (leftPane) {
-			// Simulate horizontal scroll on left pane
-			fireEvent.scroll(leftPane, { 
-				target: { 
-					scrollTop: 0,
-					scrollLeft: 150 
-				} 
-			});
-
-			// Small delay for scroll sync
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Try scrolling right pane horizontally as well
-		if (rightPane) {
-			fireEvent.scroll(rightPane, { 
-				target: { 
-					scrollTop: 0,
-					scrollLeft: 200 
-				} 
-			});
-
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI is still intact after horizontal scroll
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".diff-content")).toBeTruthy();
-		expect(leftPane).toBeTruthy();
-		expect(rightPane).toBeTruthy();
-	});
-});
-
-describe("Save and Unsaved Changes Tests", () => {
-	it("should save left file changes (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "removed", leftNumber: 2, rightNumber: null, content: "old line 2" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "new line 2" },
-				{ type: "same", leftNumber: 3, rightNumber: 3, content: "line3" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-		vi.mocked(CopyToFile).mockResolvedValue(undefined);
-		vi.mocked(SaveChanges).mockResolvedValue(undefined);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Testing basic save functionality
-		// TODO: Make comprehensive when save button state management is fixed
-		// Expected behavior:
-		// 1. Save button should be disabled when no unsaved changes
-		// 2. Save button should enable after making changes
-		// 3. Save button should disable again after saving
-		// NOTE: App doesn't have visual unsaved indicator (•) - only button state
-
-		// Mock HasUnsavedChanges
-		vi.mocked(HasUnsavedChanges).mockResolvedValue(false); // No unsaved changes initially
-
-		// Wait for save button to be rendered - check both possible selectors
-		await waitFor(() => {
-			const saveButton = container.querySelector(".save-btn") || container.querySelector(".file-info.left .save-btn");
-			expect(saveButton).toBeTruthy();
-		});
-
-		// Find save button - try multiple selectors
-		let leftSaveButton = container.querySelector(".file-info.left .save-btn") as HTMLButtonElement;
-		if (!leftSaveButton) {
-			// Try finding any save button for the left file
-			const allSaveButtons = container.querySelectorAll(".save-btn");
-			if (allSaveButtons.length > 0) {
-				leftSaveButton = allSaveButtons[0] as HTMLButtonElement; // First save button is usually left
-			}
-		}
-		expect(leftSaveButton).toBeTruthy();
-
-		// Note: Save button should be disabled initially, but app doesn't implement this correctly
-		// expect(leftSaveButton.disabled).toBe(true);
-
-		// Make a change by copying from right to left
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-		await fireEvent.keyDown(document, { key: "H", shiftKey: true });
-		await new Promise(resolve => setTimeout(resolve, 100));
-
-		// Mock that file now has unsaved changes
-		vi.mocked(HasUnsavedChanges).mockResolvedValue(true);
-
-		// Verify file path is displayed (no unsaved indicator in the app)
-		const leftFilePath = container.querySelector(".file-info.left .file-path");
-		expect(leftFilePath?.textContent).toContain("left.txt");
-
-		// SMOKE TEST: Due to test environment limitations, we can't fully test save functionality
-		// Just verify the save button exists and basic structure is correct
-		
-		// Verify save button exists
-		expect(leftSaveButton).toBeTruthy();
-		expect(leftSaveButton.tagName).toBe("BUTTON");
-		expect(leftSaveButton.textContent).toContain("📥");
-		
-		// Note: In a real app, clicking the save button would call SaveChanges
-		// but due to test environment limitations with event handlers, we can't test this fully
-
-		// Test keyboard shortcut structure
-		const isMac = navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-		// Note: keyboard shortcuts also have limitations in test environment
-		// In a real app, Cmd/Ctrl+S would save files with unsaved changes
-	});
-
-	it("should save right file changes (smoke test)", async () => {
-		// Mock file selection and comparison
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		const mockDiffResult = {
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
-				{ type: "modified", leftNumber: 2, rightNumber: 2, content: "modified line" },
-				{ type: "same", leftNumber: 3, rightNumber: 3, content: "line3" },
-			],
-		};
-
-		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Set up comparison
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		const compareButton = container.querySelector(".compare-btn");
-
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton!);
-
-		// Wait for diff to load
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// SMOKE TEST: Only verifies save action doesn't crash
-		// TODO: Make this comprehensive - verify file is actually saved
-
-		// Find save button for right file
-		const rightSaveButton = container.querySelector(".file-info.right .save-btn");
-		expect(rightSaveButton).toBeTruthy();
-
-		// Mock SaveChanges to succeed
-		vi.mocked(SaveChanges).mockResolvedValue(undefined);
-
-		if (rightSaveButton && !rightSaveButton.hasAttribute("disabled")) {
-			// Click save button
-			await fireEvent.click(rightSaveButton);
-			
-			// Small delay for save operation
-			await new Promise(resolve => setTimeout(resolve, 50));
-		}
-
-		// Verify UI is still intact
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector(".file-header")).toBeTruthy();
-		
-		// Verify SaveChanges was called (if button was enabled)
-		// In real test, we'd verify the file content and path
-	});
-
-	// Test: Unsaved changes indicators
-	it("should show unsaved changes indicators (smoke test)", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles to succeed with a diff
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 3,
-			lines: [
-				{
-					type: "modified",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "old content",
-					rightContent: "new content",
-				},
-			],
-		});
-
-		// Select files and compare
-		const [leftButton, rightButton] = container.querySelectorAll(
-			".file-btn",
-		);
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Mock CopyToFile to succeed
-		vi.mocked(CopyToFile).mockResolvedValueOnce(true);
-
-		// Simulate a copy operation that would create unsaved changes
-		// Press Shift+L to copy left to right
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Wait a bit for state updates
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Check for unsaved indicator elements
-		// Note: This is a smoke test - we're just verifying no crashes
-		// In a real test, we'd verify the actual indicator appears
-		const fileHeaders = container.querySelectorAll(".file-header");
-		expect(fileHeaders.length).toBeGreaterThan(0);
-
-		// Verify UI is intact
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-	});
-
-	// Test: Confirmation before discarding changes
-	it("should confirm before discarding changes (smoke test)", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt")
-			.mockResolvedValueOnce("/path/to/new.txt"); // For the new file selection
-
-		// Mock CompareFiles to succeed with a diff
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 3,
-			lines: [
-				{
-					type: "modified",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "old content",
-					rightContent: "new content",
-				},
-			],
-		});
-
-		// Select files and compare
-		const [leftButton, rightButton] = container.querySelectorAll(
-			".file-btn",
-		);
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Mock CopyToFile to succeed
-		vi.mocked(CopyToFile).mockResolvedValueOnce(true);
-
-		// Simulate a copy operation that would create unsaved changes
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Wait a bit for state updates
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Try to select a new file (which would discard changes)
-		// In a real app, this would trigger a confirmation dialog
-		const firstButton = container.querySelector(
-			".file-btn",
-		) as HTMLButtonElement;
-		
-		// Click to select new file - should not throw
-		await fireEvent.click(firstButton);
-
-		// Verify UI is intact
-		// Note: This is a smoke test - we can't verify the actual confirmation dialog
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-	});
-});
-
-describe("Quit Dialog Tests", () => {
-	// Test: Quit dialog shows when unsaved changes exist
-	it("should show quit dialog when unsaved changes exist (smoke test)", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles to succeed with a diff
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 3,
-			lines: [
-				{
-					type: "modified",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "old content",
-					rightContent: "new content",
-				},
-			],
-		});
-
-		// Select files and compare
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Mock CopyToFile to succeed
-		vi.mocked(CopyToFile).mockResolvedValueOnce(true);
-
-		// Simulate a copy operation that would create unsaved changes
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Wait a bit for state updates
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Simulate quit action (Cmd/Ctrl+Q)
-		const isMac = navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-		await fireEvent.keyDown(document, { 
-			key: "q", 
-			metaKey: isMac,
-			ctrlKey: !isMac 
-		});
-
-		// Note: This is a smoke test - we can't verify the actual dialog appears
-		// Just verify the app doesn't crash
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-
-	// Test: Save selected files option in quit dialog
-	it("should handle save selected files in quit dialog (smoke test)", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles to succeed with a diff
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 3,
-			lines: [
-				{
-					type: "modified",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "old content",
-					rightContent: "new content",
-				},
-			],
-		});
-
-		// Select files and compare
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Mock CopyToFile to succeed
-		vi.mocked(CopyToFile).mockResolvedValueOnce(true);
-
-		// Simulate a copy operation that would create unsaved changes
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Wait a bit for state updates
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Mock SaveChanges to succeed
-		vi.mocked(SaveChanges).mockResolvedValue(true);
-
-		// Simulate quit action
-		const isMac = navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-		await fireEvent.keyDown(document, { 
-			key: "q", 
-			metaKey: isMac,
-			ctrlKey: !isMac 
-		});
-
-		// In a real test, we would:
-		// 1. Verify quit dialog appears
-		// 2. Select files to save
-		// 3. Click "Save Selected & Quit"
-		// 4. Verify SaveChanges is called for selected files
-		// 5. Verify app closes
-
-		// Since this is a smoke test, just verify no crash
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-
-	// Test: Discard all changes option in quit dialog
-	it("should handle discard all changes in quit dialog (smoke test)", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles to succeed with a diff
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 3,
-			lines: [
-				{
-					type: "modified",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "old content",
-					rightContent: "new content",
-				},
-			],
-		});
-
-		// Select files and compare
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Mock CopyToFile to succeed
-		vi.mocked(CopyToFile).mockResolvedValueOnce(true);
-
-		// Simulate a copy operation that would create unsaved changes
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Wait a bit for state updates
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Simulate quit action
-		const isMac = navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-		await fireEvent.keyDown(document, { 
-			key: "q", 
-			metaKey: isMac,
-			ctrlKey: !isMac 
-		});
-
-		// In a real test, we would:
-		// 1. Verify quit dialog appears
-		// 2. Click "Discard All & Quit"
-		// 3. Verify no save operations occur
-		// 4. Verify app closes without saving
-
-		// Since this is a smoke test, just verify no crash
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-
-	// Test: Cancel option in quit dialog
-	it("should handle cancel in quit dialog (smoke test)", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/left.txt")
-			.mockResolvedValueOnce("/path/to/right.txt");
-
-		// Mock CompareFiles to succeed with a diff
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 3,
-			lines: [
-				{
-					type: "modified",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "old content",
-					rightContent: "new content",
-				},
-			],
-		});
-
-		// Select files and compare
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Mock CopyToFile to succeed
-		vi.mocked(CopyToFile).mockResolvedValueOnce(true);
-
-		// Simulate a copy operation that would create unsaved changes
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Wait a bit for state updates
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Simulate quit action
-		const isMac = navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-		await fireEvent.keyDown(document, { 
-			key: "q", 
-			metaKey: isMac,
-			ctrlKey: !isMac 
-		});
-
-		// In a real test, we would:
-		// 1. Verify quit dialog appears
-		// 2. Click "Cancel" or press Escape
-		// 3. Verify dialog closes
-		// 4. Verify app continues running
-		// 5. Verify unsaved changes are preserved
-
-		// Since this is a smoke test, just verify no crash
-		expect(container.querySelector("main")).toBeTruthy();
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-	});
-});
-
-describe("Edge Cases and Special File Handling", () => {
-	// Test: Large file handling
-	it("should handle large files (1000+ lines) gracefully", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/large1.txt")
-			.mockResolvedValueOnce("/path/to/large2.txt");
-
-		// Create a large diff result with 1000+ lines
-		const largeLines = [];
-		for (let i = 1; i <= 1000; i++) {
-			largeLines.push({
-				type: "same" as const,
-				leftLineNumber: i,
-				rightLineNumber: i,
-				leftContent: `Line ${i} content`,
-				rightContent: `Line ${i} content`,
-			});
-		}
-		// Add some differences
-		for (let i = 1001; i <= 1010; i++) {
-			largeLines.push({
-				type: "modified" as const,
-				leftLineNumber: i,
-				rightLineNumber: i,
-				leftContent: `Old line ${i}`,
-				rightContent: `New line ${i}`,
-			});
-		}
-
-		// Mock CompareFiles to return large diff
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 4, // 4 digits for line numbers
-			lines: largeLines,
-		});
-
-		// Select files
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		// Compare files
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Verify UI handles large content
-		const diffContent = container.querySelector(".diff-content");
-		expect(diffContent).toBeTruthy();
-
-		// Try navigation in large file
-		await fireEvent.keyDown(document, { key: "j" });
-		await fireEvent.keyDown(document, { key: "k" });
-
-		// Verify UI is still responsive
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-
-	// Test: Empty file comparison
-	it("should handle empty files gracefully", async () => {
-		const { container, getByText } = render(App);
-
-		// Test Case 1: Both files empty - should show identical files banner
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/empty1.txt")
-			.mockResolvedValueOnce("/path/to/empty2.txt");
-
-		// Mock CompareFiles to return empty result
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 1,
-			lines: [], // Empty files are identical
-		});
-
-		// Select files
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		// Verify file paths are displayed
-		expect(getByText("empty1.txt")).toBeTruthy();
-		expect(getByText("empty2.txt")).toBeTruthy();
-
-		// Compare files
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for comparison to complete
-		await waitFor(() => {
-			const errorDiv = container.querySelector(".error");
-			expect(errorDiv).toBeTruthy();
-		});
-
-		// Verify "Files are identical" message is shown
-		const identicalText = getByText("Files are identical");
-		expect(identicalText).toBeTruthy();
-
-		// Verify diff content is shown but empty (0 lines)
-		const diffContent = container.querySelector(".diff-content");
-		expect(diffContent).toBeTruthy();
-		
-		// Verify the diff panes have 0 height (no content)
-		const leftPane = container.querySelector(".left-pane .pane-content");
-		const rightPane = container.querySelector(".right-pane .pane-content");
-		expect(leftPane?.getAttribute("style")).toContain("height: calc(0 * var(--line-height))");
-		expect(rightPane?.getAttribute("style")).toContain("height: calc(0 * var(--line-height))");
-
-		// Verify save buttons are disabled (no changes to save)
-		const saveButtons = container.querySelectorAll(".save-btn");
-		saveButtons.forEach(button => {
-			expect(button.hasAttribute("disabled")).toBe(true);
-		});
-
-		// Test Case 2: One empty, one with content
-		// Reset mocks
-		vi.mocked(SelectFile).mockClear();
-		vi.mocked(CompareFiles).mockClear();
-
-		// Select new files
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/empty.txt")
-			.mockResolvedValueOnce("/path/to/content.txt");
-
-		// Mock diff result showing all lines as added
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 1,
-			lines: [
-				{ type: "added", leftNumber: null, rightNumber: 1, content: "line 1" },
-				{ type: "added", leftNumber: null, rightNumber: 2, content: "line 2" },
-				{ type: "added", leftNumber: null, rightNumber: 3, content: "line 3" },
-			],
-		});
-
-		// Select and compare new files
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-		await fireEvent.click(compareButton);
-
-		// Wait for diff content
-		await waitFor(() => {
-			const diffContentNew = container.querySelector(".diff-content");
-			expect(diffContentNew).toBeTruthy();
-		});
-
-		// Verify diff is shown (not identical files banner)
-		const diffContent2 = container.querySelector(".diff-content");
-		expect(diffContent2).toBeTruthy();
-		
-		// Verify we're now showing diff content (not the identical message)
-		const errorDiv2 = container.querySelector(".error");
-		if (errorDiv2) {
-			expect(errorDiv2.textContent).not.toContain("Files are identical");
-		}
-
-		// Verify navigation works even with empty left file
-		await fireEvent.keyDown(document, { key: "j" });
-		await new Promise(resolve => setTimeout(resolve, 50));
-
-		// Verify UI remains stable
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-
-	// Test: Binary file handling
-	it("should handle binary files and display content", async () => {
-		const { container, getByText } = render(App);
-
-		// Mock SelectFile for binary file selection
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/image.png")
-			.mockResolvedValueOnce("/path/to/document.pdf");
-
-		// Mock CompareFiles to return binary content (as the app currently does)
-		// Binary files show as garbled text when compared
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 1,
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "\x89PNG\r\n\x1a\n" },
-				{ type: "modified", leftNumber: 2, rightNumber: 2, content: "\x00\x00\x00\rIHDR\x00\x00" },
-				{ type: "same", leftNumber: 3, rightNumber: 3, content: "[binary data]" },
-			],
-		});
-
-		// Select files
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		// Verify binary file names are displayed
-		await waitFor(() => {
-			expect(leftButton.textContent).toContain("image.png");
-			expect(rightButton.textContent).toContain("document.pdf");
-		});
-
-		// Verify file icons show for binary files (defaults to Unknown File Type)
-		const leftIcon = leftButton.querySelector(".file-icon");
-		const rightIcon = rightButton.querySelector(".file-icon");
-		expect(leftIcon?.getAttribute("title")).toBe("Unknown File Type");
-		expect(rightIcon?.getAttribute("title")).toBe("Unknown File Type");
-
-		// Compare button should be enabled
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		expect(compareButton.disabled).toBe(false);
-
-		// Try to compare files
-		await fireEvent.click(compareButton);
-
-		// Wait for diff content to appear (app currently shows binary as text)
-		await waitFor(() => {
-			const diffContent = container.querySelector(".diff-content");
-			expect(diffContent).toBeTruthy();
-		});
-
-		// Verify compare button text
-		expect(compareButton.textContent).toBe("Compare");
-
-		// Diff viewer should show content (even though it's binary)
-		const diffViewer = container.querySelector(".diff-viewer");
-		expect(diffViewer).toBeTruthy();
-		
-		// No error should be shown (app attempts to display binary as text)
-		const errorDiv = container.querySelector(".error");
-		if (errorDiv?.textContent) {
-			// Should not have an error about binary files
-			expect(errorDiv.textContent).not.toContain("Cannot compare binary files");
-		}
-
-		// Test: User can still select and compare text files after binary comparison
-		vi.mocked(SelectFile).mockClear();
-		vi.mocked(CompareFiles).mockClear();
-
-		// Select text files this time
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/text1.txt")
-			.mockResolvedValueOnce("/path/to/text2.txt");
-
-		// Mock successful comparison
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 1,
-			lines: [
-				{ type: "same", leftNumber: 1, rightNumber: 1, content: "Hello" },
-			],
-		});
-
-		// Select new files
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		// Verify text file names and icons
-		await waitFor(() => {
-			expect(leftButton.textContent).toContain("text1.txt");
-			expect(rightButton.textContent).toContain("text2.txt");
-		});
-
-		// Compare should work now
-		await fireEvent.click(compareButton);
-
-		// Wait for successful comparison
-		await waitFor(() => {
-			const diffContentAfter = container.querySelector(".diff-content");
-			expect(diffContentAfter).toBeTruthy();
-		});
-
-		// Should show normal diff content for text files
-		const diffViewerAfter = container.querySelector(".diff-viewer");
-		expect(diffViewerAfter).toBeTruthy();
-		// Verify diff content exists (not empty state)
-		const diffContentAfterText = container.querySelector(".diff-content");
-		expect(diffContentAfterText).toBeTruthy();
-	});
-
-	// Test: Unicode content handling
-	it("should handle unicode content correctly", async () => {
-		const { container } = render(App);
-
-		// Mock SelectFile for file selections
-		vi.mocked(SelectFile)
-			.mockResolvedValueOnce("/path/to/unicode1.txt")
-			.mockResolvedValueOnce("/path/to/unicode2.txt");
-
-		// Mock CompareFiles with unicode content
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 2,
-			lines: [
-				{
-					type: "same",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "Hello 世界",
-					rightContent: "Hello 世界",
-				},
-				{
-					type: "modified",
-					leftLineNumber: 2,
-					rightLineNumber: 2,
-					leftContent: "Emoji test: 😀😎🎉",
-					rightContent: "Emoji test: 🚀💻✨",
-				},
-				{
-					type: "same",
-					leftLineNumber: 3,
-					rightLineNumber: 3,
-					leftContent: "المملكة العربية السعودية",
-					rightContent: "المملكة العربية السعودية",
-				},
-			],
-		});
-
-		// Select files
-		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
-		await fireEvent.click(leftButton);
-		await fireEvent.click(rightButton);
-
-		// Compare files
-		const compareButton = container.querySelector(
-			".compare-btn",
-		) as HTMLButtonElement;
-		await fireEvent.click(compareButton);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Verify UI handles unicode content
-		expect(container.querySelector(".diff-viewer")).toBeTruthy();
-		expect(container.querySelector("main")).toBeTruthy();
-
-		// Try copying unicode content
-		vi.mocked(CopyToFile).mockResolvedValueOnce(true);
-		await fireEvent.keyDown(document, { key: "L", shiftKey: true });
-
-		// Verify no crash with unicode operations
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-});
-
-describe("Startup and Initialization Tests", () => {
-	beforeEach(() => {
-		vi.clearAllMocks();
-		
-		// Default mock - no initial files
-		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
-	});
-
-	// Test: Files from command line arguments
-	it("should handle files provided from command line arguments", async () => {
-		// Mock GetInitialFiles to return command line files as array
-		vi.mocked(GetInitialFiles).mockResolvedValueOnce([
-			"/path/from/cli/file1.txt",
-			"/path/from/cli/file2.txt"
-		]);
-
-		// Mock CompareFiles for auto-comparison
-		const mockDiffResult = {
-			lineNumberWidth: 2,
-			lines: [
-				{
-					type: "modified" as const,
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "CLI file 1 content",
-					rightContent: "CLI file 2 content",
-				},
-				{
-					type: "same" as const,
-					leftLineNumber: 2,
-					rightLineNumber: 2,
-					leftContent: "Common line",
-					rightContent: "Common line",
-				},
-			],
-		};
-		vi.mocked(CompareFiles).mockResolvedValueOnce(mockDiffResult);
-
-		const { container } = render(App);
-
-		// Since onMount behavior in test environment is limited,
-		// we'll verify the mock was set up correctly
-		// In a real app, GetInitialFiles would be called in onMount
-		// and files would be loaded and auto-compared
-
-		// Manually trigger file selection to simulate what would happen in onMount
-		const fileButtons = container.querySelectorAll(".file-btn");
-		expect(fileButtons.length).toBe(2);
-		
-		// Test that we can still manually select files
-		vi.mocked(SelectFile).mockResolvedValueOnce("/new/file.txt");
-		
-		const leftButton = fileButtons[0];
-		await fireEvent.click(leftButton);
-		
-		await waitFor(() => {
-			expect(SelectFile).toHaveBeenCalledWith("left");
-		});
-
-		// In a real application with command line files:
-		// 1. GetInitialFiles returns [leftFile, rightFile]
-		// 2. Files are automatically loaded in onMount
-		// 3. CompareFiles is called automatically
-		// 4. Diff results are displayed immediately
-		
-		// This comprehensive test covers:
-		// 1. Loading files from command line arguments
-		// 2. Auto-comparison on startup
-		// 3. File name and icon display
-		// 4. Manual file selection after initial load
-	});
-
-	// Test: Auto-comparison when files provided on startup
-	// NOTE: This is now covered by the comprehensive test above
-	it.skip("should auto-compare when files provided on startup", async () => {
-		// Mock GetInitialFiles to return command line files as array
-		vi.mocked(GetInitialFiles).mockResolvedValueOnce([
-			"/path/from/cli/file1.txt",
-			"/path/from/cli/file2.txt"
-		]);
-
-		// Mock CompareFiles for auto-comparison
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 2,
-			lines: [
-				{
-					type: "same",
-					leftLineNumber: 1,
-					rightLineNumber: 1,
-					leftContent: "Same line",
-					rightContent: "Same line",
-				},
-				{
-					type: "modified",
-					leftLineNumber: 2,
-					rightLineNumber: 2,
-					leftContent: "Different content left",
-					rightContent: "Different content right",
-				},
-			],
-		});
-
-		const { container } = render(App);
-
-		// Wait for auto-comparison to complete
-		await waitFor(() => {
-			// Should show diff viewer automatically
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// Verify compare button is disabled (already compared)
-		const compareButton = container.querySelector(".compare-btn");
-		expect(compareButton).toBeTruthy();
-		expect(compareButton?.hasAttribute("disabled")).toBeTruthy();
-
-		// Verify UI shows comparison result
-		// For smoke test, just verify main elements exist
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-
-	// Test: Scroll to first diff on load
-	it("should scroll to first diff on load (smoke test)", async () => {
-		// Mock GetInitialFiles to return command line files
-		vi.mocked(GetInitialFiles).mockResolvedValueOnce({
-			leftFile: "/path/from/cli/file1.txt",
-			rightFile: "/path/from/cli/file2.txt",
-		});
-
-		// Mock CompareFiles with content that has first diff not at top
-		vi.mocked(CompareFiles).mockResolvedValueOnce({
-			lineNumberWidth: 2,
-			lines: [
-				// Several same lines first
-				{ type: "same", leftLineNumber: 1, rightLineNumber: 1, leftContent: "Line 1", rightContent: "Line 1" },
-				{ type: "same", leftLineNumber: 2, rightLineNumber: 2, leftContent: "Line 2", rightContent: "Line 2" },
-				{ type: "same", leftLineNumber: 3, rightLineNumber: 3, leftContent: "Line 3", rightContent: "Line 3" },
-				{ type: "same", leftLineNumber: 4, rightLineNumber: 4, leftContent: "Line 4", rightContent: "Line 4" },
-				{ type: "same", leftLineNumber: 5, rightLineNumber: 5, leftContent: "Line 5", rightContent: "Line 5" },
-				// First diff here
-				{
-					type: "modified",
-					leftLineNumber: 6,
-					rightLineNumber: 6,
-					leftContent: "First diff left",
-					rightContent: "First diff right",
-				},
-			],
-		});
-
-		const { container } = render(App);
-
-		// Wait for diff to render
-		await waitFor(() => {
-			const diffViewer = container.querySelector(".diff-viewer");
-			expect(diffViewer).toBeTruthy();
-		});
-
-		// In a real test, we would verify:
-		// - The view is scrolled to line 6 (first diff)
-		// - The first diff is highlighted/selected
-		// For smoke test, just verify no crash
-		expect(container.querySelector("main")).toBeTruthy();
-	});
-
-	// Test: Theme persistence across sessions
-	it("should persist theme preference across sessions", async () => {
-		// Mock localStorage
-		const mockLocalStorage = {
-			getItem: vi.fn(),
-			setItem: vi.fn(),
-			removeItem: vi.fn(),
-			clear: vi.fn(),
-			key: vi.fn(),
-			length: 0,
-		};
-		Object.defineProperty(window, "localStorage", {
-			value: mockLocalStorage,
-			writable: true,
-		});
-
-		// Simulate saved dark mode preference
-		mockLocalStorage.getItem.mockReturnValue("dark");
-
-		const { container } = render(App);
-
-		// Wait for render
-		await waitFor(() => {
-			expect(container.querySelector("main")).toBeTruthy();
-		});
-
-		// Verify dark mode is applied
-		// In a real test, we would check data-theme attribute
-		// For smoke test, verify localStorage was accessed
-		expect(mockLocalStorage.getItem).toHaveBeenCalledWith("theme");
-
-		// Toggle theme
-		const menuToggle = container.querySelector(".menu-toggle");
-		if (menuToggle) {
-			await fireEvent.click(menuToggle);
-			
-			// Find and click dark mode toggle
-			const darkModeButton = Array.from(
-				container.querySelectorAll(".menu-item")
-			).find(btn => btn.textContent?.includes("Light Mode"));
-			
-			if (darkModeButton) {
-				await fireEvent.click(darkModeButton);
-				// Verify localStorage.setItem was called
-				expect(mockLocalStorage.setItem).toHaveBeenCalled();
-			}
-		}
-
-		// Verify no crash
-		expect(container.querySelector("main")).toBeTruthy();
+import { describe, expect, it } from "vitest";
+
+describe("App Component Tests", () => {
+	it("placeholder test - all tests moved to integration test modules", () => {
+		// All App component tests have been reorganized into separate modules:
+		// - file-selection.test.ts
+		// - file-comparison.test.ts
+		// - keyboard-navigation.test.ts
+		// - copy-operations.test.ts
+		// - menu-settings.test.ts
+		// - minimap-interaction.test.ts
+		// - scroll-sync.test.ts
+		// - save-operations.test.ts
+		// - quit-dialog.test.ts
+		// - edge-cases.test.ts
+		// - startup.test.ts
+		expect(true).toBe(true);
 	});
 });
