+++ frontend/src/tests/integration/file-comparison.test.ts
@@ -1,0 +1,299 @@
+import { fireEvent, render, waitFor } from "@testing-library/svelte";
+import { beforeEach, describe, expect, it, vi } from "vitest";
+import App from "../../App.svelte";
+
+// Mock the Wails runtime
+vi.mock("../../../wailsjs/runtime/runtime.js", () => ({
+	EventsOn: vi.fn(),
+	EventsOff: vi.fn(),
+}));
+
+// Mock the Wails App functions
+vi.mock("../../../wailsjs/go/main/App.js", () => ({
+	CompareFiles: vi.fn(),
+	GetInitialFiles: vi.fn(),
+	GetMinimapVisible: vi.fn(),
+	SetMinimapVisible: vi.fn(),
+	HasUnsavedChanges: vi.fn(),
+	SaveChanges: vi.fn(),
+	CopyToFile: vi.fn(),
+	DiscardAllChanges: vi.fn(),
+	QuitWithoutSaving: vi.fn(),
+	SaveSelectedFilesAndQuit: vi.fn(),
+	SelectFile: vi.fn(),
+}));
+
+import {
+	CompareFiles,
+	GetInitialFiles,
+	GetMinimapVisible,
+	SelectFile,
+} from "../../../wailsjs/go/main/App.js";
+
+describe("App Component - File Comparison", () => {
+	beforeEach(() => {
+		vi.clearAllMocks();
+
+		// Default mocks
+		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
+		vi.mocked(GetMinimapVisible).mockResolvedValue(true);
+	});
+
+	it("should show activity indicator during comparison", async () => {
+		// Mock file selection
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		// Mock CompareFiles to have a delay to see the loading state
+		const mockDiffResult = {
+			lines: [
+				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
+				{ type: "added", leftNumber: null, rightNumber: 2, content: "added" },
+			],
+		};
+
+		vi.mocked(CompareFiles).mockImplementation(
+			() =>
+				new Promise((resolve) =>
+					setTimeout(() => resolve(mockDiffResult), 200),
+				),
+		);
+
+		const { container } = render(App);
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		// Select files
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+
+		// Click compare
+		await fireEvent.click(compareButton!);
+
+		// Should immediately show "Comparing..."
+		expect(compareButton?.textContent).toBe("Comparing...");
+		expect(compareButton).toHaveProperty("disabled", true);
+
+		// Wait for comparison to complete
+		await waitFor(
+			() => {
+				expect(compareButton?.textContent).toBe("Compare");
+				// Button should remain disabled after comparison completes
+				expect(compareButton).toHaveProperty("disabled", true);
+			},
+			{ timeout: 300 },
+		);
+	});
+
+	it("should display diff results after comparison", async () => {
+		// Mock file selection
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		// Mock CompareFiles with a simple diff result
+		const mockDiffResult = {
+			lines: [
+				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
+				{
+					type: "added",
+					leftNumber: null,
+					rightNumber: 2,
+					content: "added line",
+				},
+				{
+					type: "removed",
+					leftNumber: 2,
+					rightNumber: null,
+					content: "removed line",
+				},
+				{
+					type: "modified",
+					leftNumber: 3,
+					rightNumber: 3,
+					content: "modified line",
+				},
+				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line4" },
+			],
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
+
+		const { container } = render(App);
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		// Select files and compare
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for diff viewer to appear and comparison to complete
+		await waitFor(() => {
+			const diffViewer = container.querySelector(".diff-viewer");
+			expect(diffViewer).toBeTruthy();
+			expect(diffViewer?.classList.contains("comparing")).toBe(false);
+		});
+
+		// Check that diff content is displayed
+		const diffContent = container.querySelector(".diff-content");
+		expect(diffContent).toBeTruthy();
+
+		// Check for left and right panes
+		const leftPane = container.querySelector(".left-pane");
+		const rightPane = container.querySelector(".right-pane");
+		expect(leftPane).toBeTruthy();
+		expect(rightPane).toBeTruthy();
+
+		// Check for center gutter
+		const centerGutter = container.querySelector(".center-gutter");
+		expect(centerGutter).toBeTruthy();
+
+		// Check that content areas exist
+		const leftPaneContent = leftPane?.querySelector(".pane-content");
+		const rightPaneContent = rightPane?.querySelector(".pane-content");
+		expect(leftPaneContent).toBeTruthy();
+		expect(rightPaneContent).toBeTruthy();
+
+		// Verify that the diff viewer is showing content (not empty state)
+		const emptyState = container.querySelector(".empty-state");
+		expect(emptyState).toBeFalsy();
+	});
+
+	it("should display 'Files are identical' message for identical files", async () => {
+		// Mock file selection
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/identical.txt")
+			.mockResolvedValueOnce("/path/to/identical-copy.txt");
+
+		// Mock CompareFiles to return empty diff (identical files)
+		const mockIdenticalResult = {
+			lines: [],
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockIdenticalResult);
+
+		const { container } = render(App);
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		// Select files and compare
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for the "Files are identical" error message to appear
+		await waitFor(() => {
+			const errorDiv = container.querySelector(".error");
+			expect(errorDiv).toBeTruthy();
+			expect(errorDiv?.textContent).toBe("Files are identical");
+		});
+
+		// Verify diff viewer is shown but with no diff content
+		const diffViewer = container.querySelector(".diff-viewer");
+		expect(diffViewer).toBeTruthy();
+
+		// No diff panes should show when files are identical
+		const diffContent = container.querySelector(".diff-content");
+		expect(diffContent).toBeTruthy();
+	});
+
+	it("should display warning when comparing file to itself", async () => {
+		// Mock file selection - same file path for both
+		const samePath = "/path/to/same-file.js";
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce(samePath)
+			.mockResolvedValueOnce(samePath);
+
+		// Mock CompareFiles to return some diff result (even though files are the same)
+		const mockDiffResult = {
+			lines: [
+				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
+			],
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
+
+		const { container } = render(App);
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		// Select the same file for both sides
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+
+		// Both buttons should show the same filename
+		await waitFor(() => {
+			expect(leftButton.textContent).toContain("same-file.js");
+			expect(rightButton.textContent).toContain("same-file.js");
+		});
+
+		// Compare the files
+		await fireEvent.click(compareButton!);
+
+		// Wait for the same file warning banner to appear
+		await waitFor(() => {
+			const sameFileBanner = container.querySelector(".same-file-banner");
+			expect(sameFileBanner).toBeTruthy();
+			expect(sameFileBanner?.textContent).toContain(
+				"File same-file.js is being compared to itself",
+			);
+
+			// Check for the warning icon
+			const warningIcon = sameFileBanner?.querySelector(".warning-icon");
+			expect(warningIcon).toBeTruthy();
+			expect(warningIcon?.textContent).toBe("⚠️");
+		});
+
+		// Verify diff viewer is still shown
+		const diffViewer = container.querySelector(".diff-viewer");
+		expect(diffViewer).toBeTruthy();
+
+		// Diff content should still be shown even when comparing to itself
+		const diffContent = container.querySelector(".diff-content");
+		expect(diffContent).toBeTruthy();
+	});
+
+	it("should handle comparison errors gracefully", async () => {
+		// Mock file selection
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		// Mock CompareFiles to throw an error
+		const errorMessage = "Failed to read file: Permission denied";
+		vi.mocked(CompareFiles).mockRejectedValue(new Error(errorMessage));
+
+		const { container } = render(App);
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		// Select files and try to compare
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for the error message to appear
+		await waitFor(() => {
+			const errorDiv = container.querySelector(".error");
+			expect(errorDiv).toBeTruthy();
+			expect(errorDiv?.textContent).toContain(errorMessage);
+		});
+
+		// Compare button should be re-enabled after error
+		await waitFor(() => {
+			expect(compareButton).toHaveProperty("disabled", false);
+			expect(compareButton?.textContent).toBe("Compare");
+		});
+
+		// No diff viewer should be shown when there's an error
+		const diffViewer = container.querySelector(".diff-viewer");
+		expect(diffViewer).toBeTruthy();
+
+		// But it should show empty state, not diff content
+		const emptyState = container.querySelector(".empty-state");
+		expect(emptyState).toBeTruthy();
+		expect(emptyState?.textContent).toContain('Click "Compare Files" button');
+	});
+});
