+++ frontend/src/tests/integration/minimap-interaction.test.ts
@@ -1,0 +1,422 @@
+import { fireEvent, render, waitFor } from "@testing-library/svelte";
+import { beforeEach, describe, expect, it, vi } from "vitest";
+import App from "../../App.svelte";
+
+// Mock the Wails runtime
+vi.mock("../../../wailsjs/runtime/runtime.js", () => ({
+	EventsOn: vi.fn(),
+	EventsOff: vi.fn(),
+}));
+
+// Mock the Wails App functions
+vi.mock("../../../wailsjs/go/main/App.js", () => ({
+	CompareFiles: vi.fn(),
+	GetInitialFiles: vi.fn(),
+	GetMinimapVisible: vi.fn(),
+	SetMinimapVisible: vi.fn(),
+	HasUnsavedChanges: vi.fn(),
+	SaveChanges: vi.fn(),
+	CopyToFile: vi.fn(),
+	DiscardAllChanges: vi.fn(),
+	QuitWithoutSaving: vi.fn(),
+	SaveSelectedFilesAndQuit: vi.fn(),
+	SelectFile: vi.fn(),
+}));
+
+import {
+	CompareFiles,
+	GetInitialFiles,
+	GetMinimapVisible,
+	SelectFile,
+} from "../../../wailsjs/go/main/App.js";
+
+describe("Minimap Interaction Tests", () => {
+	beforeEach(() => {
+		vi.clearAllMocks();
+
+		// Default mocks
+		vi.mocked(GetInitialFiles).mockResolvedValue(["", ""]);
+		vi.mocked(GetMinimapVisible).mockResolvedValue(true);
+	});
+
+	it("should navigate to position when clicking on minimap (smoke test)", async () => {
+		// Mock file selection and comparison to show minimap
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		const mockDiffResult = {
+			lines: [
+				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
+				{
+					type: "added",
+					leftNumber: null,
+					rightNumber: 2,
+					content: "added line",
+				},
+				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
+				{
+					type: "removed",
+					leftNumber: 3,
+					rightNumber: null,
+					content: "removed line",
+				},
+				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
+				{
+					type: "modified",
+					leftNumber: 5,
+					rightNumber: 5,
+					content: "modified line",
+				},
+				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
+			],
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
+
+		const { container } = render(App);
+
+		// Set up comparison
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for diff to load
+		await waitFor(() => {
+			const diffContent = container.querySelector(".diff-content");
+			expect(diffContent).toBeTruthy();
+		});
+
+		// SMOKE TEST: Only verifies minimap click doesn't crash
+		// TODO: Make this comprehensive - verify actual scroll position changes
+
+		// Find the minimap
+		const minimap = container.querySelector(".minimap");
+		if (minimap) {
+			// Try clicking somewhere on the minimap
+			const minimapRect = minimap.getBoundingClientRect();
+			const clickEvent = new MouseEvent("click", {
+				bubbles: true,
+				cancelable: true,
+				clientX: minimapRect.left + minimapRect.width / 2,
+				clientY: minimapRect.top + minimapRect.height / 2,
+			});
+
+			minimap.dispatchEvent(clickEvent);
+
+			// Small delay for any scroll animation
+			await new Promise((resolve) => setTimeout(resolve, 100));
+		}
+
+		// Verify UI is still intact
+		expect(container.querySelector(".diff-viewer")).toBeTruthy();
+		expect(container.querySelector(".diff-content")).toBeTruthy();
+	});
+
+	it("should highlight current diff in minimap (smoke test)", async () => {
+		// Mock file selection and comparison to show minimap with multiple diffs
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		const mockDiffResult = {
+			lines: [
+				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
+				{
+					type: "added",
+					leftNumber: null,
+					rightNumber: 2,
+					content: "added line",
+				},
+				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
+				{
+					type: "removed",
+					leftNumber: 3,
+					rightNumber: null,
+					content: "removed line",
+				},
+				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
+				{
+					type: "modified",
+					leftNumber: 5,
+					rightNumber: 5,
+					content: "modified line",
+				},
+				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
+			],
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
+
+		const { container } = render(App);
+
+		// Set up comparison
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for diff to load
+		await waitFor(() => {
+			const diffContent = container.querySelector(".diff-content");
+			expect(diffContent).toBeTruthy();
+		});
+
+		// SMOKE TEST: Only verifies minimap renders with chunks and highlighting elements
+		// TODO: Make this comprehensive - verify specific chunks are highlighted based on current diff
+
+		// Find the minimap
+		const minimap = container.querySelector(".minimap");
+		expect(minimap).toBeTruthy();
+
+		// Check if minimap has diff chunks rendered
+		const minimapChunks = minimap?.querySelectorAll(".minimap-chunk");
+		if (minimapChunks) {
+			expect(minimapChunks.length).toBeGreaterThan(0);
+
+			// Look for highlight-related classes or elements
+			// In a real test we'd verify which chunk is highlighted
+			const highlightedElements = minimap?.querySelectorAll(
+				".highlight, .current-chunk, [class*='highlight']",
+			);
+			// Just verify the highlight mechanism exists
+			expect(highlightedElements).toBeDefined();
+		}
+
+		// Verify UI structure is intact
+		expect(container.querySelector(".diff-viewer")).toBeTruthy();
+		expect(container.querySelector(".minimap")).toBeTruthy();
+	});
+
+	it("should show viewport indicator in minimap (smoke test)", async () => {
+		// Mock file selection and comparison to show minimap
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		// Create a longer diff to ensure scrolling is possible
+		const mockDiffResult = {
+			lines: Array.from({ length: 50 }, (_, i) => ({
+				type: i % 10 === 0 ? "modified" : "same",
+				leftNumber: i + 1,
+				rightNumber: i + 1,
+				content: `line ${i + 1}`,
+			})),
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
+
+		const { container } = render(App);
+
+		// Set up comparison
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for diff to load
+		await waitFor(() => {
+			const diffContent = container.querySelector(".diff-content");
+			expect(diffContent).toBeTruthy();
+		});
+
+		// SMOKE TEST: Only verifies viewport indicator element exists
+		// TODO: Make this comprehensive - verify viewport position updates with scroll
+
+		// Find the minimap
+		const minimap = container.querySelector(".minimap");
+		expect(minimap).toBeTruthy();
+
+		// Look for viewport indicator element
+		const viewportIndicator = minimap?.querySelector(
+			".minimap-viewport, .viewport-indicator, [class*='viewport']",
+		);
+		expect(viewportIndicator).toBeTruthy();
+
+		// Simulate a scroll event (though it won't actually scroll in test environment)
+		const leftPane = container.querySelector(".diff-pane.left");
+		if (leftPane) {
+			fireEvent.scroll(leftPane, { target: { scrollTop: 100 } });
+			await new Promise((resolve) => setTimeout(resolve, 50));
+		}
+
+		// Verify UI structure is still intact after scroll attempt
+		expect(container.querySelector(".diff-viewer")).toBeTruthy();
+		expect(container.querySelector(".minimap")).toBeTruthy();
+
+		// Viewport indicator should still exist
+		const viewportAfterScroll = minimap?.querySelector(
+			".minimap-viewport, .viewport-indicator, [class*='viewport']",
+		);
+		expect(viewportAfterScroll).toBeTruthy();
+	});
+
+	it("should handle viewport drag to scroll in minimap (smoke test)", async () => {
+		// Mock file selection and comparison to show minimap
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		// Create a longer diff to ensure scrolling is possible
+		const mockDiffResult = {
+			lines: Array.from({ length: 50 }, (_, i) => ({
+				type: i % 10 === 0 ? "modified" : "same",
+				leftNumber: i + 1,
+				rightNumber: i + 1,
+				content: `line ${i + 1}`,
+			})),
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
+
+		const { container } = render(App);
+
+		// Set up comparison
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for diff to load
+		await waitFor(() => {
+			const diffContent = container.querySelector(".diff-content");
+			expect(diffContent).toBeTruthy();
+		});
+
+		// SMOKE TEST: Only verifies drag interaction doesn't crash
+		// TODO: Make this comprehensive - verify drag actually scrolls the diff view
+
+		// Find the minimap and viewport indicator
+		const minimap = container.querySelector(".minimap");
+		expect(minimap).toBeTruthy();
+
+		const viewportIndicator = minimap?.querySelector(
+			".minimap-viewport, .viewport-indicator, [class*='viewport']",
+		);
+		expect(viewportIndicator).toBeTruthy();
+
+		if (viewportIndicator) {
+			// Simulate drag interaction
+			const rect = viewportIndicator.getBoundingClientRect();
+
+			// Mouse down on viewport
+			fireEvent.mouseDown(viewportIndicator, {
+				clientX: rect.left + rect.width / 2,
+				clientY: rect.top + rect.height / 2,
+			});
+
+			// Mouse move (drag)
+			fireEvent.mouseMove(document, {
+				clientX: rect.left + rect.width / 2,
+				clientY: rect.top + rect.height / 2 + 50,
+			});
+
+			// Mouse up to end drag
+			fireEvent.mouseUp(document);
+
+			await new Promise((resolve) => setTimeout(resolve, 50));
+		}
+
+		// Verify UI structure is still intact after drag
+		expect(container.querySelector(".diff-viewer")).toBeTruthy();
+		expect(container.querySelector(".minimap")).toBeTruthy();
+		expect(viewportIndicator).toBeTruthy();
+	});
+
+	it("should show tooltip with line numbers on minimap hover (smoke test)", async () => {
+		// Mock file selection and comparison to show minimap
+		vi.mocked(SelectFile)
+			.mockResolvedValueOnce("/path/to/left.txt")
+			.mockResolvedValueOnce("/path/to/right.txt");
+
+		const mockDiffResult = {
+			lines: [
+				{ type: "same", leftNumber: 1, rightNumber: 1, content: "line1" },
+				{
+					type: "added",
+					leftNumber: null,
+					rightNumber: 2,
+					content: "added line",
+				},
+				{ type: "same", leftNumber: 2, rightNumber: 3, content: "line2" },
+				{
+					type: "removed",
+					leftNumber: 3,
+					rightNumber: null,
+					content: "removed line",
+				},
+				{ type: "same", leftNumber: 4, rightNumber: 4, content: "line3" },
+				{
+					type: "modified",
+					leftNumber: 5,
+					rightNumber: 5,
+					content: "modified line",
+				},
+				{ type: "same", leftNumber: 6, rightNumber: 6, content: "line4" },
+			],
+		};
+
+		vi.mocked(CompareFiles).mockResolvedValue(mockDiffResult);
+
+		const { container } = render(App);
+
+		// Set up comparison
+		const [leftButton, rightButton] = container.querySelectorAll(".file-btn");
+		const compareButton = container.querySelector(".compare-btn");
+
+		await fireEvent.click(leftButton);
+		await fireEvent.click(rightButton);
+		await fireEvent.click(compareButton!);
+
+		// Wait for diff to load
+		await waitFor(() => {
+			const diffContent = container.querySelector(".diff-content");
+			expect(diffContent).toBeTruthy();
+		});
+
+		// SMOKE TEST: Only verifies hover interaction doesn't crash
+		// TODO: Make this comprehensive - verify tooltip appears with correct line numbers
+
+		// Find the minimap
+		const minimap = container.querySelector(".minimap");
+		expect(minimap).toBeTruthy();
+
+		if (minimap) {
+			// Simulate hover over minimap
+			const rect = minimap.getBoundingClientRect();
+
+			// Mouse enter minimap
+			fireEvent.mouseEnter(minimap);
+
+			// Mouse move within minimap
+			fireEvent.mouseMove(minimap, {
+				clientX: rect.left + rect.width / 2,
+				clientY: rect.top + rect.height / 2,
+			});
+
+			await new Promise((resolve) => setTimeout(resolve, 100));
+
+			// Look for tooltip element (might be in document body or near minimap)
+			const tooltip = document.querySelector(
+				".minimap-tooltip, .tooltip, [role='tooltip']",
+			);
+			// Just verify tooltip mechanism exists, actual content would require proper initialization
+
+			// Mouse leave minimap
+			fireEvent.mouseLeave(minimap);
+		}
+
+		// Verify UI structure is still intact
+		expect(container.querySelector(".diff-viewer")).toBeTruthy();
+		expect(container.querySelector(".minimap")).toBeTruthy();
+	});
+});
